Food systems update pt. 1
Food is no longer picked up by collision, instead it is now picked up by hitting right mouse button near it. Food can now be spit out, coonsuming the food. If it does not land inside the home zone, the food is re-instantiated on collision. Food is simply dropped in the home zone with right mouse button as before.